requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1+cu117
torchvision==0.14.1+cu117
torchaudio==0.13.1
gym==0.23.1
numpy==1.21.6
pygame==2.3.0
tqdm==4.65.0
pyyaml==6.0
matplotlib==3.5.3
opencv-python==4.7.0.72
pandas==1.5.3
seaborn==0.12.2
shapely==1.8.5
ephem==4.1.4
joblib==1.2.0
cpprb==10.7.0
pycocotools==2.0.6
moviepy==1.0.3
scikit-image==0.19.3
setGPU